+++ ecuacion-splib-web-jpa/src/main/java/jp/ecuacion/splib/web/jpa/service/SplibGeneralJpaService.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package jp.ecuacion.splib.web.service;
+package jp.ecuacion.splib.web.jpa.service;
 
 import jp.ecuacion.lib.jpa.entity.AbstractEntity;
+import jp.ecuacion.splib.web.service.SplibGeneralService;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional(rollbackFor = Exception.class)
